+++ admin/grouping/featuregroup.php
@@ -28,6 +28,12 @@ function leafext_help_featuregroup() {
 	__( 'Group elements and dynamically add/remove from map', 'extensions-leaflet-map' ) .
 	'.</p>';
 
+	if ( is_singular() || is_archive() ) {
+		$parentref = get_site_url() . '/doku/parentgroup/';
+	} else {
+		$parentref = '?page=' . LEAFEXT_PLUGIN_SETTINGS . '&tab=parentgroup';
+	}
+
 	$text = $text . '<p>' .
 	sprintf(
 		/* translators: %s are shortcodes. */
@@ -37,7 +43,7 @@ function leafext_help_featuregroup() {
 		),
 		'<ul><li> <code>leaflet-optiongroup</code>',
 		'</li><li> <code>leaflet-featuregroup</code>',
-		'</li><li> <code>leaflet-parentgroup</code>'
+		'</li><li> <a href="' . $parentref . '"><code>leaflet-parentgroup</code></a>'
 	)
 	. '</li></ul></p>';
 
@@ -276,24 +282,24 @@ function leafext_help_featuregroup() {
 	<pre' . $codestyle . '><code' . $codestyle . '>&#091;leaflet-optiongroup option="..." values="...,...,others,unknown" groups="...,...,Other elements,Unknown elements"]
 &#091;leaflet-featuregroup property="..." values="...,...,others,unknown" groups="...,...,Other elements,Unknown elements"]</code></pre>';
 
-	$text = $text . '<h3>leaflet-parentgroup</h3><p>
-<pre' . $codestyle . '><code' . $codestyle . '>&#091;leaflet-parentgroup parent=... childs=... expandall=... collapseall=...]</code></pre>';
-
-	$text = $text . '<p>' . __(
-		'Display groups in a tree view.',
-		'extensions-leaflet-map'
-	);
-	$text = $text . '</p><ul>
-	<li><code>parent</code> - ' . __( 'Name of the parent', 'extensions-leaflet-map' ) . '</li>
-	<li><code>childs</code> - ' . sprintf(
-		/* translators: %s is an option. */
-		__( 'child names - comma separated group names (the same as in %s above)', 'extensions-leaflet-map' ),
-		'<code>groups</code>'
-	) . '. ' .
-	__( 'If you have html tags there, you can omit these here', 'extensions-leaflet-map' ) .
-	'.</li>';
-	$text = $text . '<li><code>expandall</code>, <code>collapseall</code> - ' . __( 'Text for an entry in control that expands or collapses the tree. If empty (default), no entry is created. The specification in the first command applies.', 'extensions-leaflet-map' ) . '</li>';
-	$text = $text . '</ul>';
+	//  $text = $text . '<h3>leaflet-parentgroup</h3><p>
+	// <pre' . $codestyle . '><code' . $codestyle . '>&#091;leaflet-parentgroup parent=... childs=... expandall=... collapseall=...]</code></pre>';
+	//
+	//  $text = $text . '<p>' . __(
+	//      'Display groups in a tree view.',
+	//      'extensions-leaflet-map'
+	//  );
+	//  $text = $text . '</p><ul>
+	//  <li><code>parent</code> - ' . __( 'Name of the parent', 'extensions-leaflet-map' ) . '</li>
+	//  <li><code>childs</code> - ' . sprintf(
+	//      /* translators: %s is an option. */
+	//      __( 'child names - comma separated group names (the same as in %s above)', 'extensions-leaflet-map' ),
+	//      '<code>groups</code>'
+	//  ) . '. ' .
+	//  __( 'If you have html tags there, you can omit these here', 'extensions-leaflet-map' ) .
+	//  '.</li>';
+	//  $text = $text . '<li><code>expandall</code>, <code>collapseall</code> - ' . __( 'Text for an entry in control that expands or collapses the tree. If empty (default), no entry is created. The specification in the first command applies.', 'extensions-leaflet-map' ) . '</li>';
+	//  $text = $text . '</ul>';
 	$text = $text . '<h3>' . __( 'Group Control', 'extensions-leaflet-map' ) . '</h3>' .
 	'<p>' .
 	'<pre' . $codestyle . '><code' . $codestyle . '>&#091;leaflet-optiongroup ... position=topleft|topright|bottomleft|bottomright collapsed=true|false]